+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 0.1.15
+- fixed: CognitoUser.getCachedDeviceKeyAndPassword as a Future<void>
+- removed flutter_test dev dependency
+
+## 0.1.14
+- added: CognitoUser.getDeviceKey method
+- added documentation for client secret
+
+## 0.1.13
+- changed: CognitoUser.updateAttributes and CognitoUser.deleteAttributes return Future<bool>
+- added: CognitoUser.getMFAOptions method
+
+## 0.1.12+3
+- fix: type list<dynamic> is not a subtype of type list<int>
+
 ## 0.1.12+2
 - fix: getUserAttributes strict casting issue